requirements-gpu.txt
tensorflow-gpu==2.3.0rc0
opencv-python==4.2.0
lxml
tqdm
absl-py
matplotlib
easydict
pillow
